publisher/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sse-starlette>=2.0.0
playwright>=1.45.0
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv
Pillow>=10.0.0
aiofiles>=23.0
